fixed some

+++ src/coil-sensor.cpp
@@ -15,7 +15,7 @@ int16_t CoilSensor_Read()
     uint16_t leftVal = analogRead(sensor_left_pin);
     uint16_t rightVal = analogRead(sensor_right_pin);
 
-    int16_t difference = leftVal - rightVal;
+    int16_t difference = (int16_t)leftVal - (int16_t)rightVal;
 
     return difference;
 }
